+++ src/filesystem/filesystem.cpp
@@ -5,11 +5,11 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
+#include "filesystem.hpp"
+
 #include <unistd.h>
 
 #include <climits>
-
-#include <zephir/filesystem.hpp>
 
 bool zephir::filesystem::Exists(const std::string &name) {
   return (access(name.c_str(), F_OK) != -1);
